chore: use OperationResultScreenContent for FL modals
